feat: move exchange helpers to utils
