+++ README.MD
@@ -22,4 +22,4 @@ Run `bytes32(uint(uint160(<V2_ASSET_ADDRESS>)) << 96) | bytes32(<SUB_ID>)`
 1. `yarn build` and `yarn link` inside of `derive-sentio-utils`
 2. `rm -rf node_modules` and `yarn cache clean` and `yarn link @derivefinance/derive-sentio-utils` and `yarn install` in subrepo that uses utils
 
-If adding new graphQL schemas, make sure to export in `index.ts` and re-run `yarn build` and `yarn sentio build`
+If adding new graphQL schemas, make sure to export in `index.ts` and re-run BOTH `yarn sentio build` then `yarn build` (order matters).
