+++ derive-sentio-utils/src/index.ts
@@ -2,10 +2,10 @@ import { IntegratorSeason, getCurrentSeason } from "./integratorSeasons.js";
 
 import { estimateBlockNumberAtDate } from "./crosschainBlocks.js";
 
-
+export * as contracts from './types/eth/index.js';
 export * as constants from './constants.js';
 export * as vaults from './vaults/index.js';
-export * as v2Subgraph from './v2/v2Subgraph.js';
+export * as v2 from './v2/index.js';
 
 export * as schemas from './schema/store.js';
