+++ derive-sentio-utils/src/v2/constants.ts
@@ -1,0 +1,2 @@
+export const DERIVE_V2_SUBACCOUNTS_ADDRESS = "0xE7603DF191D699d8BD9891b821347dbAb889E5a5";
+export const DERIVE_V2_MATCHING_ADDRESS = "0xeB8d770ec18DB98Db922E9D83260A585b9F0DeAD";
