--- etherfi-integration/abis/eth/DeriveVaultToken.json
@@ -1,21 +1,0 @@
-[
-    {
-        "type": "function",
-        "name": "getSharesValue",
-        "inputs": [
-            {
-                "name": "numShares",
-                "type": "uint256",
-                "internalType": "uint256"
-            }
-        ],
-        "outputs": [
-            {
-                "name": "",
-                "type": "uint256",
-                "internalType": "uint256"
-            }
-        ],
-        "stateMutability": "view"
-    }
-]
